requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
sqlalchemy==2.0.36
pydantic==2.9.2
pandas==2.2.3
openpyxl==3.1.5
python-dateutil==2.9.0
plotly==5.24.1
python-dotenv==1.0.1
python-multipart
httpx==0.27.0

